backend/requirements.txt
flask
google-auth
google-api-python-client
mammoth
requests